+++ example/lib/src/cache/ttl_5_minutes_cache_policy.dart
@@ -1,4 +1,3 @@
-
 import 'package:zuraffa/zuraffa.dart';
 
 /// Auto-generated cache policy