+++ test/plugins/provider/provider_plugin_test.dart
@@ -53,9 +53,7 @@ void main() {
       isTrue,
     );
     expect(
-      content1.contains(
-        "import 'package:app/src/domain/entities/listing/listing.dart';",
-      ),
+      content1.contains("import '../../domain/entities/listing/listing.dart';"),
       isTrue,
     );
     expect(content1.contains('listing_?'), isFalse);